refactor(FOROME-725): split  both modals into one
